\name{fpop2D-package}
\alias{fpop2D-package}
\alias{fpop2D}
\docType{package}
\title{
  Functional Pruning Optimal Partitioning algorithm of  bivariate time-series.
}
\description{
  fpop2D is an R package written in Rcpp C++ and developed to change-point detection in bivariate time-series.
}
\details{
  The package implements the Functional Pruning Optimal Partitioning algorithm with two types of pruning: "intersection of sets" and "difference of intersection and union of sets". 
}
\author{
Liudmila Pishchagina, ludmilashabarshova1991\@gmail.com.

Maintainer: Liudmila Pishchagina <ludmilashabarshova1991\@gmail.com.com>
}
\references{
  
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
