% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpop2D.R
\name{fpop2D}
\alias{fpop2D}
\title{Functional pruning optimal partitioning for 2d time-series (fpop2D)}
\usage{
fpop2D(df, data1, data2, penalty, weights = NULL, testMode = FALSE)
}
\arguments{
\item{df}{dataframe of size n x 2}

\item{data1}{vector of data 1st dimension}

\item{data2}{vector of data 1st dimension}

\item{penalty}{a positive number to penalize the introduction of changepoints}

\item{weights}{vector of weights (positive numbers), same size as data}

\item{testMode}{boolean. False by default. Used to debug the code}
}
\value{
a fpop2d object = (changepoints, parameters, globalCost)
\describe{
\item{\code{changepoints}}{is the vector of changepoints (we give the last element of each segment)}
\item{\code{parameters}}{is the vector of successive parameters of each segment}
\item{\code{globalCost}}{is a number equal to the total loss: the minimal cost for the optimization problem with all penalty values excluded}
 }
}
\description{
Functional pruning optimal partitioning for 2d time-series
}
