% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpop2D.R
\name{plotFPOP2D}
\alias{plotFPOP2D}
\title{plotFPOP2D}
\usage{
plotFPOP2D(data, changepoints, means1, means2)
}
\arguments{
\item{data}{matrix of data of dimension 2 x n.}

\item{changepoints}{a vector of increasing changepoint indices (last element is always n!).}

\item{means1}{vector of successive means1 for data1.}

\item{means2}{vector of successive means2 for data2.}

\item{n}{number of data point.}
}
\value{
Plot of data with a given values of means and changepoints.
}
\description{
Plot of data with a  values of means and changepoints.
}
\examples{
plotFPOP2D(data = matrix(c(1,0,1,0,1,0,1,0,2,3,2,3,2,3,2,3), nrow = 2), changepoints = c(5, 8), means1 = c(1, 2), means2 = c(0,3))
}
