% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FPOP2D}
\alias{FPOP2D}
\title{FPOP2D}
\usage{
FPOP2D(data1, data2, penalty, type)
}
\arguments{
\item{data1}{is a vector of data1(a univariate time series).}

\item{data2}{is a vector of data2(a univariate time series).}

\item{penalty}{is a value of penalty (a non-negative real number).}

\item{type}{is a value defining the  type of pruning (0 = PELT, 1 = FPOP(intersection of sets), 2 = FPOP(intersection of set \ union of set ).}
}
\value{
a list of 4 elements  = (changepoints, means1, means2, globalCost).                    
 
\describe{
\item{\code{changepoints}}{is the vector of changepoints.}
\item{\code{means1}}{is the vector of successive means for data1.}
\item{\code{means2}}{is the vector of successive means for data2.}
\item{\code{globalCost}}{is a number equal to the global cost.}
}                                                                                                                                                                             #
}
\description{
Detecting changepoints using the functional pruning optimal partitioning method (fpop) in bivariate time series.
}
\examples{
FPOP2D (data1 = c(0,0,0,1,1,1), data2 = c(2,2,2,0,0,0), penalty = 2*log(6),  type = 1) 
}
